sort, filter, page, group

+++ ContosoUniversity/ContosoUniversity/Controllers/StudentsController.cs
@@ -20,13 +20,51 @@ namespace ContosoUniversity.Controllers
         }
 
         // GET: Students
-        public async Task<IActionResult> Index()
-        {
-            return View(await _context.Students.ToListAsync());
-        }
+		public async Task<IActionResult> Index(string sortOrder,
+	string currentFilter,
+	string searchString,
+	int? page)
+		{
+			ViewData["CurrentSort"] = sortOrder;
+			ViewData["NameSortParm"] = String.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
+			ViewData["DateSortParm"] = sortOrder == "Date" ? "date_desc" : "Date";
+			if (searchString != null)
+			{
+				page = 1;
+			}
+			else
+			{
+				searchString = currentFilter;
+			}
+			ViewData["CurrentFilter"] = searchString;
+			var students = from s in _context.Students
+						   select s;
+			if (!String.IsNullOrEmpty(searchString))
+			{
+				students = students.Where(s => s.LastName.Contains(searchString)
+									   || s.FirstMidName.Contains(searchString));
+			}
+			switch (sortOrder)
+			{
+				case "name_desc":
+					students = students.OrderByDescending(s => s.LastName);
+					break;
+				case "Date":
+					students = students.OrderBy(s => s.EnrollmentDate);
+					break;
+				case "date_desc":
+					students = students.OrderByDescending(s => s.EnrollmentDate);
+					break;
+				default:
+					students = students.OrderBy(s => s.LastName);
+					break;
+			}
+			int pageSize = 3;
+			return View(await PaginatedList<Student>.CreateAsync(students.AsNoTracking(), page ?? 1, pageSize));
+		}
 
-        // GET: Students/Details/5
-        public async Task<IActionResult> Details(int? id)
+		// GET: Students/Details/5
+		public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
             {
@@ -127,7 +165,7 @@ namespace ContosoUniversity.Controllers
 			return View(studentToUpdate);
 		}
 		// GET: Students/Delete/5
-		public async Task<IActionResult> Delete(int? id)
+		public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -135,13 +173,22 @@ namespace ContosoUniversity.Controllers
             }
 
             var student = await _context.Students
+				.AsNoTracking()
                 .SingleOrDefaultAsync(m => m.ID == id);
             if (student == null)
             {
                 return NotFound();
             }
 
-            return View(student);
+			if (saveChangesError.GetValueOrDefault())
+			{
+				ViewData["ErrorMessage"] =
+					"Delete failed. Try again, and if the problem persists " +
+					"see your system administrator.";
+			}
+
+
+			return View(student);
         }
 
         // POST: Students/Delete/5
@@ -149,11 +196,25 @@ namespace ContosoUniversity.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var student = await _context.Students.SingleOrDefaultAsync(m => m.ID == id);
-            _context.Students.Remove(student);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
+            var student = await _context.Students
+				.AsNoTracking()
+				.SingleOrDefaultAsync(m => m.ID == id);
+			if (student == null)
+			{
+				return RedirectToAction(nameof(Index));
+			}
+			try
+			{
+				_context.Students.Remove(student);
+				await _context.SaveChangesAsync();
+				return RedirectToAction(nameof(Index));
+			}
+			catch (DbUpdateException /* ex */)
+			{
+				//Log the error (uncomment ex variable name and write a log.)
+				return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+			}
+		}
 
         private bool StudentExists(int id)
         {
